[python] move parse input functions to input_parser.py
